fix(tests): silence cross-compiler warnings in unit tests
- async/fs_tests.cpp: rename `event_loop& loop` lambda params to `ev`
  so they stop hiding loop_fixture's `loop` member (MSVC C4458).
- async/process_tests.cpp: wrap the pipe-chunk helper in
  `#ifndef _WIN32` so clang-cl does not flag it as an unused function.
- async/stream_tests.cpp: cast message length to int on Windows where
  `::send` takes int rather than size_t.
- deco/runtime.cc: suppress C5046 around the thread_local destructor
  registration and explicitly narrow the trace-size capture to
  unsigned (C4267).

+++ tests/unit/async/stream_tests.cpp
@@ -489,7 +489,7 @@ TEST_CASE(accept_and_read) {
     ASSERT_EQ(::connect(client_fd, reinterpret_cast<sockaddr*>(&addr), sizeof(addr)), 0);
 
     const std::string message = "kotatsu-tcp";
-    ASSERT_EQ(::send(client_fd, message.data(), message.size(), 0),
+    ASSERT_EQ(::send(client_fd, message.data(), static_cast<int>(message.size()), 0),
               static_cast<ssize_t>(message.size()));
     close_socket(client_fd);
